Improve Channels organization
