+++ src/main/java/org/example/scalingstream/channels/AbstractChannel.kt
@@ -1,5 +1,8 @@
 package org.example.scalingstream.channels
 
-abstract class AbstractChannel<Type>(override val id: ChannelID, val channelArgs: ChannelArgs) : Channel<Type> {
-    abstract val type: String
+abstract class AbstractChannel<Type>(
+    override val id: ChannelID,
+    val channelArgs: ChannelArgs
+) : Channel<Type> {
+
 }
