+++ src/main/java/org/example/scalingstream/channels/Channel.kt
@@ -2,14 +2,15 @@ package org.example.scalingstream.channels
 
 import java.util.*
 
-typealias ChannelID = Pair<UUID, UUID>
+data class ChannelID(val src: UUID, val dst: UUID)
 
 interface Channel<Type> {
     val id: ChannelID
+    val type: String
 
     fun getChannelReader(): InputChannel<Type>
 
-    fun getChannelWriter() : OutputChannel<Type>
+    fun getChannelWriter(): OutputChannel<Type>
 
     fun destroy()
 }
