--- src/main/java/org/example/scalingstream/channels/DataChannel.kt
@@ -1,11 +1,0 @@
-package org.example.scalingstream.channels
-
-import java.time.Instant
-
-typealias Record<Type> = Pair<Instant?, List<Type>>
-
-abstract class DataChannel(var id: ChannelID) {
-
-    abstract fun connect()
-
-}
