+++ src/main/java/org/example/scalingstream/channels/OutputChannel.kt
@@ -1,10 +1,9 @@
 package org.example.scalingstream.channels
 
 import java.time.Instant
-import java.util.*
 
 
-abstract class OutputChannel<Type>(id: ChannelID) : DataChannel(id) {
+abstract class OutputChannel<Type>(override val id: ChannelID) : ChannelIO {
     abstract fun put(recordBatch: Record<Type>)
     abstract fun flush()
     abstract fun close(timestamp: Instant?)
